LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/module/referrer-no-referrer.sub-expected.txt

CONSOLE MESSAGE: Blocked access to external URL http://www1.localhost:8800/html/semantics/scripting-1/the-script-element/module/resources/referrer-checker.py?name=remote
CONSOLE MESSAGE: Blocked access to external URL http://www1.localhost:8800/html/semantics/scripting-1/the-script-element/module/resources/import-referrer-checker.sub.js?name=remote_remote
CONSOLE MESSAGE: Blocked access to external URL http://www1.localhost:8800/html/semantics/scripting-1/the-script-element/module/resources/import-same-origin-referrer-checker-from-remote-origin.sub.js?name=remote_same
CONSOLE MESSAGE: Blocked access to external URL http://www1.localhost:8800/html/semantics/scripting-1/the-script-element/module/resources/referrer-checker.py?name=same_remote

Harness Error (TIMEOUT), message = null
